🚀 Comprehensive GitHub Actions fixes and optimizations
- Fixed LTO compatibility issues on macOS by disabling LTO for clang on Darwin
- Fixed compiler warnings in json_schema_generator.c, json_parser_simd.c, and cpu_features.c
- Enhanced test scripts with better error handling and build verification
- Optimized GitHub Actions workflows with timeouts and caching
- Added comprehensive performance testing script with multi-tier benchmarking
- Added platform compatibility checker for cross-platform validation
- Added comprehensive test runner for all test types
- Fixed Memory Safety Testing workflow with proper AddressSanitizer flags
- Enhanced CodeQL Security Analysis with better configuration and timeouts
- Added performance report generation and artifact uploads

All GitHub Actions workflows should now run successfully without timeouts or build failures.

+++ c-lib/src/cpu_features.c
@@ -213,7 +213,7 @@ static void detect_cpu_cores(cpu_features_t* features) {
             }
         }
         fclose(fp);
-        features->num_cores = physical_cores > 0 ? physical_cores : features->num_logical_cores;
+        features->num_cores = physical_cores > 0 ? (unsigned int)physical_cores : features->num_logical_cores;
     } else {
         features->num_cores = features->num_logical_cores;
     }
